refactor: lib/csv/parsers.ts をフォーマット別ファイルに分割 (#82)
498行の parsers.ts を7ファイルに分割:
- types.ts: 型定義 + 共有ユーティリティ
- wise.ts / revolut.ts / rakuten.ts / smbc.ts / generic.ts: 各パーサー
- index.ts: re-export + フォーマット判定 + ディスパッチャー

振る舞い変更なし、既存テスト278件全通過
